requirements.txt
django==3.1.6 
djangorestframework==3.11.2
drf-yasg==1.17.1
django-cors-headers==3.3.0
django-extensions==2.1.9